[minorfix,test] kstring_t init macro, typo in release action, pileup test
[minorfix]

fix KS_INITIALIZE not working with static linking

fix typo in release github action

fix arg value printing for --rm-invar-sites helppage

[test]

fix pileup testing and update for updated rng function

[chore]

add better check for number of quality score ranges

+++ io.h
@@ -144,7 +144,7 @@ struct argStruct {
     /// qs_bins[i][1] = end of i-th range
     /// qs_bins[i][2] = qs value to assign for i-th range
     uint8_t** qs_bins;
-    int n_qs_bins;
+    uint8_t n_qs_bins;
 
     // functions:
     double* read_depthsFile(void);
